create LogHoursPopupWindowForm component and add styling
